+++ frontend/src/views/Pages/AccountOverview.vue
@@ -24,9 +24,9 @@
   </div>
 </template>
 <script>
-import GddStatus from '../KontoOverview/GddStatus.vue'
-import GddSend from '../KontoOverview/GddSend.vue'
-import GddTable from '../KontoOverview/GddTable.vue'
+import GddStatus from './AccountOverview/GddStatus.vue'
+import GddSend from './AccountOverview/GddSend.vue'
+import GddTable from './AccountOverview/GddTable.vue'
 
 export default {
   name: 'Overview',
